RISCV/shutdown: Implement machine_{halt,restart}()
SBI has an API for shutdown so wire it up.  However, the spec does allow the
call not to be implemented, so we have to cope with sbi_shutdown() returning.

There is a reboot-capable SBI extention, but in the short term route
machine_restart() into machine_halt().

Then, use use machine_halt() rather than an infinite loop at the end of
start_xen().  This avoids the Qemu smoke test needing to wait for the full
timeout in order to succeed.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: Jan Beulich <jbeulich@suse.com>
Reviewed-by: Oleksii Kurochko <oleksii.kurochko@gmail.com>